requirements.txt
requests
shapely>=1.6.4
cloudpickle
numpy
pandas
requests-mock
deprecated